requirements.txt
mss
pillow
pyserial
pywin32
random